move config into separate file
